+++ src/components/GoodList.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const GoodList = ({ goods }) => (
+  <ul>
+    {goods.map(good => (
+      <li data-cy="Good" key={good}>{good}</li>
+    ))}
+  </ul>
+);
